+++ console/settings.py
@@ -2,10 +2,8 @@ import os
 import logging
 import json
 
-from sdx.common.logger_config import logger_initial_config
-
-
-logger_initial_config(service_name='sdx-console')
+LOGGING_LEVEL = logging.getLevelName(os.getenv('LOGGING_LEVEL', 'DEBUG'))
+LOGGING_FORMAT = "%(asctime)s.%(msecs)06dZ|%(levelname)s: sdx-console: %(message)s"
 
 logger = logging.getLogger(__name__)