+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+  - Remove sdx-common logging and move logging code back into the repo
   - Add Cloudfoundry deployment files
 
 ### 1.5.1 2017-10-25